requirements.txt
ultralytics
streamlit
pillow
torch
